Small changes for fix bugs.
git-svn-id: https://svn.java.net/svn/docsources~svn/trunk@1282 3b25252b-017b-4ae5-8f61-fa87a855a72e

+++ DocSources/src/main/java/org/medici/docsources/service/peoplebase/PeopleBaseService.java
@@ -288,7 +288,7 @@ public interface PeopleBaseService {
 	 * @return
 	 * @throws ApplicationThrowable
 	 */
-	public Map<Integer, Integer> findNumbersOfDocumentsRelated(List<Integer> personIds) throws ApplicationThrowable;
+	public Map<Integer, Long> findNumbersOfDocumentsRelated(List<Integer> personIds) throws ApplicationThrowable;
 	
 	/**
 	 *
